+++ src/modules/budget/application/Data/Service/IBudgetImportService.cs
@@ -9,9 +9,9 @@ namespace Intive.Patronage2023.Modules.Budget.Application.Data.Service;
 public interface IBudgetImportService
 {
 	/// <summary>
-	/// Import method.
+	/// Imports budget data from the provided file.
 	/// </summary>
-	/// <param name="file">IFormFile.</param>
+	/// <param name="file">The file containing the budget data to import.</param>
 	/// <returns>A <see cref="Task{TResult}"/> representing the result of the asynchronous operation.</returns>
 	Task<GetImportResult> Import(IFormFile file);
 }
